Zmiana cen na double

+++ FakTest/DodanieProduktu.xaml.cs
@@ -22,7 +22,7 @@ namespace FakTest
         Controler _controler;
 
         public string _nazwa { get; set; }
-        public int _cena { get; set; }
+        public double _cena { get; set; }
         public int _stawkaVat { get; set; }
 
         public DodanieProduktu(Controler controler)
@@ -63,9 +63,9 @@ namespace FakTest
             produkt.typ = "brak";
             produkt.netto = linia.cena.ToString() + "zł";
             produkt.stawka = linia.VAT.ToString() + "%";
-            int temp = (linia.cena * linia.VAT / 100);
+            double temp = (linia.cena * linia.VAT / 100);
             produkt.podatek = temp.ToString() + " zł";
-            int temp2 = linia.cena + temp;
+            double temp2 = linia.cena + temp;
             produkt.brutto = temp2.ToString() + " zł";
             DataGridProduktow.Items.Add(produkt);
         }
@@ -81,9 +81,9 @@ namespace FakTest
                 produkt.typ = "brak";
                 produkt.netto = linia.cena.ToString() + "zł";
                 produkt.stawka = linia.VAT.ToString() + "%";
-                int temp = (linia.cena * linia.VAT / 100);
+                double temp = (linia.cena * linia.VAT / 100);
                 produkt.podatek = temp.ToString() + " zł";
-                int temp2 = linia.cena + temp;
+                double temp2 = linia.cena + temp;
                 produkt.brutto = temp2.ToString() + " zł";
                 DataGridProduktow.Items.Add(produkt);
             }
